Moved stacked-area-chart page  to group 5 folder
