+++ src/pages/index.ts
@@ -1,6 +1,6 @@
 export { default as Overview } from "./Overview";
 export { default as Feedback } from "./Feedback";
-export { default as StackedAreaPage } from "./StackedAreaPage";
+export { default as StackedAreaPage } from "./Group5/StackedAreaPage";
 export { default as SunburstPage } from "./SunburstPage";
 export { default as TreemapPage } from "./TreeMapPage";
 export { default as Group11Page } from "./Group11";
